requirements.txt
mysql-connector-python
mwparserfromhell
pywikibot
requests
unidecode
